[REFACTOR] KeyBoardHeight 함수 클래스화 하기  (#324)
